Adjusted for changes in fast API, removed macro API
