+++ packages/sources/spectral-macro-score/schemas/env.json
@@ -4,45 +4,15 @@
   "required": ["API_KEY", "RPC_URL", "NFC_REGISTRY_ADDRESS"],
   "type": "object",
   "properties": {
-    "BASE_URL_MACRO_API": {
-      "type": "string",
-      "format": "uri",
-      "description": "An API key that can be obtained from the data provider's dashboard",
-      "required": true
-    },
     "BASE_URL_FAST_API": {
       "type": "string",
       "format": "uri",
-      "description": "Ethereum RPC URL",
+      "description": "Fast API URL",
       "required": true
-    },
-    "MACRO_API_KEY": {
-      "type": "string",
-      "format": "uri"
     },
     "FAST_API_KEY": {
       "type": "string",
       "format": "uri"
-    },
-    "INFURA_URL": {
-      "type": "string",
-      "format": "uri"
-    },
-    "INFURA_API_KEY": {
-      "type": "string",
-      "format": "uri"
-    },
-    "NFC_REGISTRY_ADDRESS": {
-      "type": "string",
-      "description": "Address of the NFC contract",
-      "format": "web3 address",
-      "required": true
-    },
-    "API_ENDPOINT": {
-      "type": "string",
-      "description": "MACRO Score API Endpoint",
-      "format": "web3 address",
-      "default": "https://xzff24vr3m.execute-api.us-east-2.amazonaws.com/default"
     }
   },
   "allOf": [
