+++ packages/sources/spectral-macro-score/src/endpoint/index.ts
@@ -1,2 +1,2 @@
-export * as MacroScoreAPI from './macro-score-api'
+export * as MacroScoreAPI from './fast-api-score'
 export * as ExtraDataAPI from './extra-data'
